+++ SummerEngine/SummerEngine/Texture.cpp
@@ -37,12 +37,12 @@ const Texture::ResourceType Texture::GetType() const
 
 float Texture::GetHeight()
 {
-	return GetDesc2D()->Height;
+	return (float)GetDesc2D()->Height;
 }
 
 float Texture::GetWidth()
 {
-	return GetDesc2D()->Width;
+	return (float)GetDesc2D()->Width;
 }
 
 D3D11_TEXTURE2D_DESC* Texture::GetDesc2D()
@@ -51,5 +51,5 @@ D3D11_TEXTURE2D_DESC* Texture::GetDesc2D()
 	ID3D11Resource* t_Resource;
 	GetTextureView()->GetResource(&t_Resource);
 	reinterpret_cast<ID3D11Texture2D*>(t_Resource)->GetDesc(&t_Desc);
-	return &t_Desc;
+	return &t_Desc; // TODO  this is a temp var? 
 }
